Move ListDef and MapDef into their respective modules
