+++ facet-core/src/types/map.rs
@@ -1,5 +1,73 @@
 use crate::opaque::{Opaque, OpaqueConst, OpaqueUninit};
 
+use super::Shape;
+
+/// Fields for map types
+#[derive(Clone, Copy, PartialEq, Eq, Hash, Debug)]
+#[repr(C)]
+#[non_exhaustive]
+pub struct MapDef {
+    /// vtable for interacting with the map
+    pub vtable: &'static MapVTable,
+    /// shape of the keys in the map
+    pub k: &'static Shape,
+    /// shape of the values in the map
+    pub v: &'static Shape,
+}
+
+impl MapDef {
+    /// Returns a builder for MapDef
+    pub const fn builder() -> MapDefBuilder {
+        MapDefBuilder::new()
+    }
+}
+
+/// Builder for MapDef
+pub struct MapDefBuilder {
+    vtable: Option<&'static MapVTable>,
+    k: Option<&'static Shape>,
+    v: Option<&'static Shape>,
+}
+
+impl MapDefBuilder {
+    /// Creates a new MapDefBuilder
+    #[allow(clippy::new_without_default)]
+    pub const fn new() -> Self {
+        Self {
+            vtable: None,
+            k: None,
+            v: None,
+        }
+    }
+
+    /// Sets the vtable for the MapDef
+    pub const fn vtable(mut self, vtable: &'static MapVTable) -> Self {
+        self.vtable = Some(vtable);
+        self
+    }
+
+    /// Sets the key shape for the MapDef
+    pub const fn k(mut self, k: &'static Shape) -> Self {
+        self.k = Some(k);
+        self
+    }
+
+    /// Sets the value shape for the MapDef
+    pub const fn v(mut self, v: &'static Shape) -> Self {
+        self.v = Some(v);
+        self
+    }
+
+    /// Builds the MapDef
+    pub const fn build(self) -> MapDef {
+        MapDef {
+            vtable: self.vtable.unwrap(),
+            k: self.k.unwrap(),
+            v: self.v.unwrap(),
+        }
+    }
+}
+
 /// Initialize a map in place with a given capacity
 ///
 /// # Safety
